INS-1979: move genesis bootstrap code to 'bootstrap' dir
