+++ scripts/generate_insolar_configs.go
@@ -26,11 +26,11 @@ import (
 	"strings"
 	"time"
 
+	"github.com/insolar/insolar/bootstrap/genesis"
 	pulsewatcher "github.com/insolar/insolar/cmd/pulsewatcher/config"
 	"github.com/insolar/insolar/configuration"
-	"github.com/insolar/insolar/genesis"
 	"github.com/spf13/cobra"
-	yaml "gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v2"
 )
 
 func check(msg string, err error) {
